refactor(oauth): shuffling of location of api type declarations in oauth plugin
`globals.ts` should keep only global constants used within a package, so let's
locate common types in `api.ts` or `types.ts`.

Also, moved one orphaned util function to a more proper place.
