+++ packages/oauth/src/utils/index.ts
@@ -3,4 +3,3 @@ export * from './fetch'
 export * from './https'
 export * from './strings'
 export * from './urls'
-export * from './utils'
