--- packages/oauth/src/utils/utils.ts
@@ -1,3 +1,0 @@
-export function relToAbsUrl(relativeUrl: string): string {
-  return new URL(relativeUrl, window.location.origin).href
-}
